# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Job Vacancy API
  version: 0.6.0
  description: |
    Shared contract for the web (React) and mobile (Flutter) clients.

    **Realtime.** POST /ws/ticket, then open `wss://…/api/v1/ws?ticket=…`. Server events:
    `ready`, `message.new`, `message.read`, `message.deleted`, `typing`, `presence`,
    `presence.snapshot`, `notification.new`, `pong`, `error` — each `{"type", "data"}`.
    Client frames: `{"type":"ping"}`, `{"type":"typing","conversation_id"}`,
    `{"type":"read","conversation_id","message_id"}`, `{"type":"watch_presence","user_ids":[…]}`.
    Events are not buffered: after reconnecting, fetch `/conversations/{id}/messages?after=<last id>`.

    **Uploads.** POST /files → multipart POST to `upload.url` with every `upload.fields`
    entry followed by the `file` field → POST /files/{id}/complete → use the file id.

    **Envelope.** Success: `{"data": ...}` (lists add `"meta"`). Failure:
    `{"error": {"code", "message", "fields"?, "retry_after"?}}`. `code` is stable; clients
    translate it into uz / uz-Cyrl / ru / en. `fields` maps a JSON field to the failing
    rule (`required`, `email`, `min=8`, `password`, `oneof=...`).

    **Deadlines.** Every request runs under a server deadline (8 s; PDF export 25 s;
    the WebSocket is exempt) and each database statement under 5 s. When one is exceeded
    the API answers `503` with code `timeout` and `Retry-After: 1`; the call is safe to
...
    limits of their own.

...

...
    it rendered.

...

...
servers:
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
...
    CompanyRef: { name: company, in: path, required: true, schema: { type: string }, description: uuid or slug }
    IfNoneMatch:
      name: If-None-Match
      in: header
      schema: { type: string }
      description: ETag of a copy the client has; unchanged content answers 304 with no body.
...
  schemas:
    Locale: { type: string, enum: [uz, uz-Cyrl, ru, en] }
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [invalid_credentials] }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
            retry_after: { type: integer, description: seconds }
...
    UserResponse:
      type: object
      properties: { data: { $ref: "#/components/schemas/User" } }
...
    CompanyInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        industry_id: { type: integer, description: category id }
        size: { type: string, enum: ["1-10", "11-50", "51-200", "201-500", "501-1000", "1000+"] }
        website: { type: string, description: "scheme optional: najottalim.uz → https://najottalim.uz" }
        email: { type: string, format: email }
        phone: { type: string }
        region_id: { type: integer, description: top-level region }
        address: { type: string, maxLength: 300 }
        about: { type: string, maxLength: 5000 }
        founded_year: { type: integer, minimum: 1800, maximum: 2100 }
    Company:
      allOf:
        - $ref: "#/components/schemas/CompanyInput"
        - type: object
          required: [id, slug, name, logo_url, cover_url, verified, created_at, about]
          properties:
            id: { type: string, format: uuid }
            slug: { type: string }
            logo_url: { type: [string, "null"] }
            cover_url: { type: [string, "null"] }
            verified: { type: boolean }
            created_at: { type: string, format: date-time }
            open_vacancies: { type: integer, description: public profile only }
            my_role: { type: string, enum: [owner, admin, recruiter], description: "/me/companies only" }
            status: { type: string, enum: [active, blocked, deleted], description: "/me/companies only; blocked companies can't be managed" }
...
    AdminCompany:
      allOf:
        - $ref: "#/components/schemas/Company"
        - type: object
          required: [status, owner_id, owner_name, owner_email, open_vacancies]
          properties:
            status: { type: string, enum: [active, blocked, deleted] }
            owner_id: { type: string, format: uuid }
            owner_name: { type: string }
            owner_email: { type: [string, "null"] }
            open_vacancies: { type: integer }
...
    ReportReason: { type: string, enum: [spam, fraud, offensive, discrimination, misleading, duplicate, other] }
...
    AdminReport:
      type: object
      required: [id, object_type, reason, comment, status, created_at, resolved_at, resolved_by, resolution_note, reporter, vacancy, open_reports]
      properties:
        id: { type: string, format: uuid }
        object_type: { type: string, enum: [vacancy] }
        reason: { $ref: "#/components/schemas/ReportReason" }
        comment: { type: string }
        status: { type: string, enum: [open, resolved, dismissed] }
        created_at: { type: string, format: date-time }
        resolved_at: { type: [string, "null"], format: date-time }
        resolved_by: { type: [string, "null"], format: uuid }
        resolution_note: { type: string }
        reporter:
          type: [object, "null"]
          properties: { id: { type: string, format: uuid }, full_name: { type: string }, email: { type: [string, "null"] } }
        vacancy:
          type: [object, "null"]
          properties:
            id: { type: string, format: uuid }
            title: { type: string }
            slug: { type: string }
            status: { $ref: "#/components/schemas/VacancyStatus" }
            company: { type: object, properties: { id: { type: string, format: uuid }, name: { type: string }, slug: { type: string } } }
        open_reports: { type: integer, description: "open reports on the same vacancy (open queue only)" }
...
    VacancyInput:
      type: object
      required: [title, description, category_id, region_id, employment_type, work_format, experience, schedule]
      properties:
        title: { type: string, minLength: 3, maxLength: 150 }
        description: { type: string, minLength: 30, maxLength: 10000, description: Markdown }
        category_id: { type: integer }
        region_id: { type: integer, description: top-level region }
        district_id: { type: integer, description: must belong to region_id }
        address: { type: string, maxLength: 300 }
        salary_min: { type: integer, format: int64 }
        salary_max: { type: integer, format: int64 }
        currency: { type: string, enum: [UZS, USD], default: UZS }
        employment_type: { $ref: "#/components/schemas/EmploymentType" }
        work_format: { $ref: "#/components/schemas/WorkFormat" }
        experience: { $ref: "#/components/schemas/Experience" }
        schedule: { $ref: "#/components/schemas/Schedule" }
        skills: { type: array, maxItems: 20, items: { type: string, maxLength: 50 }, description: free-text names; unknown ones are created }
    EmploymentType: { type: string, enum: [full_time, part_time, project, internship, volunteer] }
    WorkFormat: { type: string, enum: [office, remote, hybrid] }
    Experience: { type: string, enum: [none, 1_3, 3_6, 6_plus] }
    Schedule: { type: string, enum: [full_day, shift, flexible, rotation] }
    VacancyStatus: { type: string, enum: [draft, moderation, published, rejected, archived, expired] }
...
    VacancyResponse:
      type: object
      properties: { data: { $ref: "#/components/schemas/VacancyDetail" } }
    VacancyPage:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/VacancyCard" } }
        meta:
          type: object
          properties:
            next_cursor: { type: [string, "null"], description: "pass as ?cursor= for the next page" }
            total: { type: integer, description: "first page of /vacancies only; capped at 1000" }
            total_capped: { type: boolean, description: "true → show \"1000+\"" }
            fuzzy: { type: boolean, description: results are typo-tolerant matches }
    ResumeInput:
      type: object
      required: [title]
      description: The whole resume; PUT replaces every section.
      properties:
        title: { type: string, minLength: 2, maxLength: 150, description: desired position }
        about: { type: string, maxLength: 5000 }
        category_id: { type: integer }
        region_id: { type: integer }
        relocate: { type: boolean }
        desired_salary: { type: integer, format: int64 }
        currency: { type: string, enum: [UZS, USD] }
        employment_types: { type: array, items: { $ref: "#/components/schemas/EmploymentType" } }
        work_formats: { type: array, items: { $ref: "#/components/schemas/WorkFormat" } }
        visibility: { $ref: "#/components/schemas/ResumeVisibility" }
        experiences: { type: array, maxItems: 20, description: most recent first, items: { $ref: "#/components/schemas/ResumeExperience" } }
        educations: { type: array, maxItems: 10, items: { $ref: "#/components/schemas/ResumeEducation" } }
        skills: { type: array, maxItems: 30, items: { type: string } }
        languages: { type: array, maxItems: 10, items: { $ref: "#/components/schemas/ResumeLanguage" } }
...
    ResumeCard:
      type: object
      required: [id, title, person, category_id, region_id, relocate, experience_months, desired_salary, last_job, skills, updated_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        person: { type: object, properties: { id: { type: string, format: uuid }, full_name: { type: string }, avatar_url: { type: [string, "null"] } } }
        category_id: { type: [integer, "null"] }
        region_id: { type: [integer, "null"] }
        relocate: { type: boolean }
        experience_months: { type: integer, description: overlapping jobs counted once }
        desired_salary: { type: [object, "null"], properties: { amount: { type: integer }, currency: { type: string } } }
        last_job: { type: [object, "null"], properties: { position: { type: string }, company: { type: string }, current: { type: boolean } } }
        skills: { type: array, items: { $ref: "#/components/schemas/Skill" } }
        visibility: { $ref: "#/components/schemas/ResumeVisibility" }
        updated_at: { type: string, format: date-time }
...
    ApplicationStatus: { type: string, enum: [sent, viewed, invited, interview, hired, rejected, withdrawn] }
...
  headers:
    ETag: { description: Validator of the response body (weak), for If-None-Match, schema: { type: string } }
    XCache: { description: "HIT | STALE | MISS: how the cached response was served", schema: { type: string, enum: [HIT, STALE, MISS] } }
  responses:
    NotModified:
      description: Not modified (the If-None-Match ETag is current); no body
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    User:
      description: Current user
      content: { application/json: { schema: { $ref: "#/components/schemas/UserResponse" } } }
    NoContent: { description: Done }
    Resume:
      description: Resume
      content: { application/json: { schema: { $ref: "#/components/schemas/ResumeResponse" } } }
...
    Vacancy:
      description: Vacancy
      content: { application/json: { schema: { $ref: "#/components/schemas/VacancyResponse" } } }
    VacancyPage:
      description: Page of vacancies
      content: { application/json: { schema: { $ref: "#/components/schemas/VacancyPage" } } }
...
    CachedVacancyPage:
      description: Page of vacancies (cached)
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
        X-Cache: { $ref: "#/components/headers/XCache" }
      content: { application/json: { schema: { $ref: "#/components/schemas/VacancyPage" } } }
    Company:
      description: Company
      content:
        application/json:
          schema: { type: object, properties: { data: { $ref: "#/components/schemas/Company" } } }
paths:
...
  /me:
    get:
      tags: [me]
      security: [{ bearer: [] }]
      summary: Current user
      responses: { "200": { $ref: "#/components/responses/User" } }
    patch:
      tags: [me]
      security: [{ bearer: [] }]
      summary: Update name, language or presence visibility
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name: { type: string, minLength: 2, maxLength: 100 }
                locale: { $ref: "#/components/schemas/Locale" }
                hide_online: { type: boolean, description: "hide the online status from everyone (TZ SEC-05); applies to live watchers at once" }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      tags: [me]
      security: [{ bearer: [] }]
      summary: Delete the account (TZ FN-03)
      description: |
        Confirm with `password`; accounts without a password (Google-only) send a fresh
        `google_id_token` or call within 10 minutes of signing in (else 403 reauth_required).
        Refused with 409 `ownership_transfer_required` (fields.companies = comma-separated
        slugs) while the user owns a company that has other members: transfer it first
        (POST /companies/{company}/owner). A company the user is the only member of closes
        with the account. Personal data is removed (name, e-mail, phone, avatar, uploads,
        resumes; applications stay with employers, shown with an empty full_name), every
        session is revoked and the e-mail can register again. Clears the refresh cookie.
        At most 10 attempts per 15 minutes (429 rate_limited).
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string, maxLength: 72 }
                google_id_token: { type: string }
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
...
  /me/companies:
    get:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Companies the current user belongs to (with my_role)
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/Company" } } } }
...
  /companies:
    get:
      tags: [companies]
      summary: Company directory (verified first, then by open vacancies), 24 per page
      description: |
        Keyset pages (TZ BE-03), never OFFSET. Page either by number (`page`, for numbered
        pagination: it seeks from a cached page anchor) or by `cursor` (from
        `meta.next_cursor`, takes precedence over `page`). `meta.total` and
        `meta.page_count` come from page anchors cached for ~3 minutes, so they (and page
        boundaries) may lag new companies or ranking changes by that long. A `page` past
        the last one returns an empty `data`. `open_vacancies` is kept by a database
        trigger on every vacancy status change.
      parameters:
        - { name: q, in: query, description: "Name contains (case-insensitive, max 100 chars)", schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: cursor, in: query, description: "meta.next_cursor of the previous page", schema: { type: string } }
      responses:
        "200":
          description: Page
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Company" } }
                  meta:
                    type: object
                    required: [page, next_page, next_cursor, total, page_count]
                    properties:
                      page: { type: [integer, "null"], description: "Requested page number; null when paging by cursor" }
                      next_page: { type: [integer, "null"] }
                      next_cursor: { type: [string, "null"] }
                      total: { type: integer, description: "Matching active companies" }
                      page_count: { type: integer }
        "400": { $ref: "#/components/responses/Error" }
    post:
      tags: [companies]
      security: [{ bearer: [] }]
      summary: Create a company (employer with a verified email or phone); caller becomes owner
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CompanyInput" } } } }
      responses:
        "201": { $ref: "#/components/responses/Company" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /companies/{company}/vacancies:
    parameters: [{ $ref: "#/components/parameters/CompanyRef" }]
    get:
      tags: [vacancies]
      security: [{ bearer: [] }]
      summary: Employer dashboard — all of the company's vacancies, newest first
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/VacancyStatus" } }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { $ref: "#/components/responses/VacancyPage" } }
    post:
      tags: [vacancies]
      security: [{ bearer: [] }]
      summary: Create a draft vacancy (any member)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/VacancyInput" } } } }
      responses:
        "201": { $ref: "#/components/responses/Vacancy" }
        "422": { $ref: "#/components/responses/Error" }
  /vacancies:
    get:
      tags: [vacancies]
      summary: Search and list published vacancies (keyset pagination)
      description: |
        `q` matches across scripts and languages: "dasturchi", "дастурчи" and "Dasturchilar"
        find the same vacancies; Russian inflections ("продажам") and Uzbek suffixes are
        handled by prefix-stemming. Title, skills, company, category and region names are all
        searchable (weighted in that order). If nothing matches exactly, a typo-tolerant title
        search runs and `meta.fuzzy` is true ("showing similar results").

        Ordering: `relevance` (default with `q`) or `newest` (default without `q`).
        Multi-value filters accept repeated params or comma lists (work_format=remote,hybrid).
        A parent category_id includes its subcategories. Results are cached ~30 s (a
        publish or edit refreshes them on the next request); `meta.total` up to 2 minutes.
      parameters:
        - { $ref: "#/components/parameters/IfNoneMatch" }
        - { name: q, in: query, schema: { type: string, maxLength: 200 } }
        - { name: sort, in: query, schema: { type: string, enum: [relevance, newest] } }
        - { name: category_id, in: query, schema: { type: integer } }
        - { name: region_id, in: query, schema: { type: integer } }
        - { name: district_id, in: query, schema: { type: integer } }
        - { name: company_id, in: query, schema: { type: string, format: uuid } }
        - { name: employment_type, in: query, schema: { type: string } }
        - { name: work_format, in: query, schema: { type: string } }
        - { name: experience, in: query, schema: { type: string } }
        - { name: schedule, in: query, schema: { type: string } }
        - { name: salary_from, in: query, schema: { type: integer } }
        - { name: currency, in: query, schema: { type: string, enum: [UZS, USD], default: UZS } }
        - { name: with_salary, in: query, schema: { type: boolean } }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { $ref: "#/components/responses/CachedVacancyPage" }
        "304": { $ref: "#/components/responses/NotModified" }
        "422": { $ref: "#/components/responses/Error" }
...
  /admin/companies:
    get:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Companies of any status, newest first
      parameters:
        - { name: q, in: query, schema: { type: string }, description: part of the name }
        - { name: status, in: query, schema: { type: string, enum: [active, blocked, deleted] } }
        - { name: verified, in: query, schema: { type: boolean } }
        - { $ref: "#/components/parameters/Cursor" }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Page
          content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/AdminCompany" } }, meta: { type: object, properties: { next_cursor: { type: [string, "null"] } } } } } } }
...
  /admin/reports:
    get:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Reports queue — open ones oldest first, resolved/dismissed newest first
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [open, resolved, dismissed], default: open } }
        - { name: vacancy_id, in: query, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/Cursor" }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Page
          content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/AdminReport" } }, meta: { type: object, properties: { next_cursor: { type: [string, "null"] } } } } } } }
...
  /resumes:
    get:
      tags: [resumes]
      security: [{ bearer: [] }]
      summary: Candidate search (employers). Public resumes only; no contacts in results.
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [relevance, updated] } }
        - { name: category_id, in: query, schema: { type: integer } }
        - { name: region_id, in: query, schema: { type: integer } }
        - { name: with_relocate, in: query, schema: { type: boolean }, description: also candidates willing to relocate }
        - { name: experience, in: query, schema: { type: string }, description: "none,1_3,3_6,6_plus (comma list)" }
        - { name: salary_to, in: query, schema: { type: integer } }
        - { name: currency, in: query, schema: { type: string, enum: [UZS, USD] } }
        - { name: employment_type, in: query, schema: { type: string } }
        - { name: work_format, in: query, schema: { type: string } }
        - { name: language, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200":
          description: Page of candidates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/ResumeCard" } }
                  meta: { type: object }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: [resumes]
      security: [{ bearer: [] }]
      summary: Create a resume (seekers, max 5)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ResumeInput" } } } }
      responses:
        "201": { $ref: "#/components/responses/Resume" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /me/resumes:
    get:
      tags: [resumes]
      security: [{ bearer: [] }]
      summary: My resumes
      responses: { "200": { description: Resumes } }
...
  /me/applications:
    get:
      tags: [applications]
      security: [{ bearer: [] }]
      summary: My applications with vacancy cards
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/ApplicationStatus" } }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: Page } }
...
  /files:
    post:
      tags: [files]
      security: [{ bearer: [] }]
      summary: Start an upload (returns a presigned POST; size and type are enforced by storage)
      description: |
        Limits: avatar / company_logo — jpeg, png, webp ≤ 5 MB (public URL);
        chat_image — jpeg, png, webp, gif ≤ 10 MB; chat_voice — webm, ogg, m4a (audio/mp4), mp3, aac ≤ 5 MB;
        chat_file — pdf, doc(x), xls(x), ppt(x), txt, zip, rar, jpeg, png ≤ 25 MB (private, signed URLs).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [purpose, content_type, size]
              properties:
                purpose: { type: string, enum: [avatar, company_logo, chat_image, chat_file, chat_voice] }
                content_type: { type: string }
                size: { type: integer }
                name: { type: string }
                meta: { type: object, additionalProperties: true, description: "{duration_ms} for voice, {width,height} for images" }
      responses: { "201": { description: "{file, upload: {url, method, fields}}" }, "422": { $ref: "#/components/responses/Error" } }
...
  /ws:
    get:
      tags: [realtime]
      summary: WebSocket endpoint
      parameters: [{ name: ticket, in: query, required: true, schema: { type: string } }]
      responses: { "101": { description: Switching protocols }, "401": { $ref: "#/components/responses/Error" } }
...
  /conversations:
    get:
      tags: [chat]
      security: [{ bearer: [] }]
      summary: My conversations, most recent first, with last message and unread count
      description: |
        Keyset pages (`meta.next_cursor`). `unread` is the number of undeleted messages
        from others after the caller's read position (capped at 100), kept up to date on
        every send, read and delete (TZ BE-04). A participant who never opened a
        conversation counts its whole history.
      parameters: [{ $ref: "#/components/parameters/Cursor" }, { $ref: "#/components/parameters/Limit" }]
      responses: { "200": { description: Page }, "400": { $ref: "#/components/responses/Error" } }
...